rename to trace categories and make it possible to add your own
