+++ lib/core/engineDelegate/chromeDelegate.js
@@ -5,7 +5,7 @@ const Promise = require('bluebird'),
   webdriver = require('selenium-webdriver'),
   harBuilder = require('../../support/harBuilder'),
   perflogParser = require('../../support/chromePerflogParser'),
-  timeLineParser = require('../../support/timeLineParser');
+  traceCategoriesParser = require('../../support/traceCategoriesParser');
 
 const CHROME_NAME_AND_VERSION_JS = `return (function() {
   var match = navigator.userAgent.match(/(Chrom(e|ium))\\/(([0-9]+\\.?)*)/);
@@ -22,10 +22,10 @@ const CHROME_NAME_AND_VERSION_JS = `return (function() {
 const PAGE_TITLE_JS = 'return document.title;';
 
 class ChromeDelegate {
-  constructor({skipHar = false, experimental = {}}) {
+  constructor({skipHar = false, experimental = {}, chrome = {}}) {
     this.skipHar = skipHar;
     this.logPerfEntries = !!experimental.dumpChromePerflog;
-    this.dumpTimeLineLog = !!experimental.dumpTimeLineLog;
+    this.dumpTraceCategoriesLog = !!chrome.dumpTraceCategoriesLog;
   }
 
   onStartRun() {
@@ -59,9 +59,9 @@ class ChromeDelegate {
         }
       })
       .tap((events) => {
-        if (this.dumpTimeLineLog) {
-          const cleaned = timeLineParser.parse(events);
-          results.extraJson['timeline-' + index + '.json'] = cleaned;
+        if (this.dumpTraceCategoriesLog) {
+          const cleaned = traceCategoriesParser.parse(events);
+          results.extraJson['trace-' + index + '.json'] = cleaned;
         }
       })
       .then((events) => perflogParser.harFromEvents(events))
