+++ lib/core/webdriver/chrome.js
@@ -2,6 +2,7 @@
 
 let webdriver = require('selenium-webdriver'),
   chrome = require('selenium-webdriver/chrome'),
+  util = require('../../support/util'),
   chromedriver = require('alto-saxophone').chromedriver;
 
 let hasConfiguredChromeDriverService = false;
@@ -34,7 +35,7 @@ const defaultChromeOptions = [
   '--ignore-certificate-errors'
 ];
 
-const timelineTraceCategories = [
+let timelineTraceCategories = [
     '-*',
     'toplevel',
     'blink.console',
@@ -77,9 +78,14 @@ module.exports.configureBuilder = function(builder, options) {
 
   chromeOptions.setLoggingPrefs(logPrefs);
   const perfLogConf = {enableNetwork: true, enablePage: true};
-  if (options.experimental && options.experimental.dumpTimeLineLog) {
+  if (chromeConfig.addTraceCategory) {
+    timelineTraceCategories.push.apply(timelineTraceCategories, util.toArray(chromeConfig.addTraceCategory));
+  }
+
+  if (chromeConfig.dumpTraceCategoriesLog) {
     perfLogConf.traceCategories = timelineTraceCategories.join(',');
   }
+
   chromeOptions.setPerfLoggingPrefs(perfLogConf);
 
   if (options.userAgent) {
