+++ lib/support/cli.js
@@ -117,6 +117,15 @@ module.exports.parseCommandLine = function parseCommandLine() {
       describe: 'Choose which device to use. If you do not set it, random device will be used.',
       group: 'chrome'
     })
+    .option('chrome.dumpTraceCategoriesLog', {
+      type: 'boolean',
+      describe: 'Dump Chromes traceCategories log to disk.',
+      group: 'chrome'
+    })
+    .option('chrome.addTraceCategory', {
+      describe: 'Add extra trace category.',
+      group: 'chrome'
+    })
     .option('firefox.binaryPath', {
       describe: 'Path to custom Firefox binary (e.g. Firefox Nightly). ' +
         'On OS X, the path should be to the binary inside the app bundle, ' +
@@ -140,11 +149,6 @@ module.exports.parseCommandLine = function parseCommandLine() {
     .option('experimental.dumpChromePerflog', {
       type: 'boolean',
       describe: 'Dump Chromes performance log with Network.* events to disk.',
-      group: 'experimental'
-    })
-    .option('experimental.dumpTimeLineLog', {
-      type: 'boolean',
-      describe: 'Dump Chromes timeline log to disk.',
       group: 'experimental'
     })
     .option('video', {
