+++ lib/support/traceCategoriesParser.js
